0608_0030
做了大地图UI和交互逻辑

+++ Assets/Scripts/Data/LocationData.cs
@@ -11,12 +11,12 @@ public class LocationData
     public string name;
     // 地点类型
     public LocationTypeEnum type;
-    // 可通行的目的地列表（key=目标地点ID, value=路径长度）
-    public Dictionary<int, float> connections;
+    // 可通行的目的地列表（key=目标地点ID, value=路径长度/时段数）
+    public Dictionary<int, int> connections;
     // 进入该地点后的顶级场景ID
     public int top_scene_id;
 
-    public LocationData(int id, string name, LocationTypeEnum type, Dictionary<int, float> connections, int top_scene_id = -1)
+    public LocationData(int id, string name, LocationTypeEnum type, Dictionary<int, int> connections, int top_scene_id = -1)
     {
         this.id = id;
         this.name = name;
@@ -28,12 +28,12 @@ public class LocationData
     /// <summary>
     /// 获取前往某地的路径长度，若不可达返回 -1
     /// </summary>
-    public float GetConnectionLength(int target_id)
+    public int GetConnectionLength(int target_id)
     {
         if (connections.ContainsKey(target_id))
         {
             return connections[target_id];
         }
-        return -1f;
+        return -1;
     }
 }
